Add files via upload
Briefly summarize The Gaming Room client and their software requirements. Who was the client? What type of software did they want you to design?
The client is the Client was a game company called the Gaming Room, they wanted us to make a web-base verson of their game "Draw It Or Loose it". The idea was to make the game mulptie platformer, able to function 
on a diverse range of operating systems. They wanted to be multiplayer, with players being able to complete on different teams. They also wanted each player to have different IDs to make differenate them.

What did you do particularly well in developing this documentation?
I did well in the summary, I was able to explaine the overall plan and what is needed to be done.

What about the process of working through a design document did you find helpful when developing the code?
It make things much simpler. With a general plan already layout it is easily to figure out what to do and what to research. 

If you could choose one part of your work on these documents to revise, what would you pick? How would you improve it?
I should have gone into detial on the "Development Requirements", giving more information and thought on each operating system. 

How did you interpret the user’s needs and implement them into your software design? Why is it so important to consider the user’s needs when designing?
Putting himself in the user's perpesctive, the user would want things to be simple and less complicated. The user's needs must be consider, for otherwise the project weren't be succesful. 
The user will want diverse options, and be able to choose their own teams. They also want to be able to socialized with their fellow players. The user's needs are importent when 
trying to create a succesful product. 

How did you approach designing software? What techniques or strategies would you use in the future to analyze and design a similar software application?
I approach software design like approaching how to build a house, it is better to have a blue print before laying the foundation. But there needs to be
room for changes in plan and be not scare in adding new details and features to make it better. In the future I will use charts and write down what needs
to be done before starting, like what classes and functions I would need to code. I would also parpare to know what to do when things go wrong.